fix(shopping): migrate Product.price from Double to BigDecimal
- Change Product entity price field from Double to BigDecimal for
  precise monetary calculations
- Add Jakarta Validation to ProductCreateRequest, ProductUpdateRequest
  (@NotBlank, @NotNull, @Positive, @Min)
- Update all DTOs and documents: ProductResponse, ProductWithReviewsResponse,
  ProductDocument, ProductSearchResult, AdminProductRequest
- Remove unnecessary BigDecimal.valueOf() wrappers in CartServiceImpl,
  TimeDealProduct, TimeDealResponse
- Add V14 Flyway migration: ALTER products.price to DECIMAL(19,2)

+++ services/shopping-service/src/main/java/com/portal/universe/shoppingservice/cart/service/CartServiceImpl.java
@@ -16,8 +16,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
-
 /**
  * 장바구니 관리 서비스 구현체입니다.
  */
@@ -60,7 +58,7 @@ public class CartServiceImpl implements CartService {
                             cart.addItem(
                                     request.productId(),
                                     product.getName(),
-                                    BigDecimal.valueOf(product.getPrice()),
+                                    product.getPrice(),
                                     request.quantity()
                             );
                         }
